docs: Adds basic documentation using mkdocs
I had an hour of work left and was unlikely to be able to concentrate deeply on work so used the time productively to
knock up the documentation for AFMReader using [mkdocs](https://www.mkdocs.org/).

Its pretty basic but once merged pushes to the `main` branch of files under `docs/` will trigger the documentation to be
rebuilt. It includes the API docs auto-generated from docstrings and the ability to easily switch between versions.

This is a prelude to switching TopoStats and [AFM-SPM.github.io](https://github.com/AFM-SPM/AFM-SPM.github.io) over to
`mkdocs`.