feature: Returns just dictionary when loading .topostats
Closes #109

Rather than return the HDF5 object loaded from `.topostats` as a dictionary as the third item in a tuple along with the
`image` and `pixel_to_nm_scaling` (both of which are extracted from the loaded data anyway) just the dictionary is
returned. This matches the structure of objects created by TopoStats and how they are saved to HDF5 files.