tests: make topostats tests flexible to key name
We are going to switch from having the `topostats_file_version` as the name for the metadata that `.topostats` files
were generated with to using the `topostats_version` as key which will align with the release tag and make it easier for
work to be reproducible.

This small commit adds the `version_key` variable to the `tests/test_topostats.py::test_load_topostats()` test so that
the key against which tests are made will work with old versions and future versions.