bug: decode numpy arrays of strings
I found whilst working through TopoStats refactoring that the inclusion of the `config` as an attribute to `TopoStats`
objects resulted in some of the values, which are lists of strings, being converted to Numpy arrays.

These could not be decoded direction and `item[()].decode("utf-8")` failed with an `AttributeError` stating that `numpy.ndarray
does not have attribute decode`.

The solution proposed here is to capture this error and if `item[()]` is an instance of `np.ndarray` to iterate over the
list decoding each item in turn.  The typing is explicitly ignored because we want it as a list rather than a
dictionary.

Test included, and whilst it passes seems a bit light but does mirror the scenario encountered (had to use `group_path`
as the higher level of nesting).