feature: handle newer topostats file versions
Restructuring of TopoStats to define classes means the `TopoStats` object now holds the `topostats_version` rather than
`topostats_file_version`.

This commit allows both to be handled and switches to using
[`packging.version`](https://packaging.pypa.io/en/stable/version.html) to do so which ensures a more consistent approach
to comparing version numbers.

Yet to write a test for working with newer `.topostats` where `topostats_version >= 2.3.2` as work is still on-going but
parameterised test is in place for when work is complete.